+++ src/tds/vstrbuild.c
@@ -45,7 +45,7 @@ struct string_linked_list
  */
 
 static char *
-norm_fmt(const char *fmt, int fmtlen)
+norm_fmt(const char *fmt, ptrdiff_t fmtlen)
 {
 	char *newfmt;
 	char *cp;
